src/requirements.txt
distribute==0.6.31
pycrypto==2.6
